refactor: Updated TooltipTour, useTour & LibraryPage after rebasing remote changes.
